requirements.txt
attrs==17.2.0
Automat==0.6.0
constantly==15.1.0
hyperlink==17.2.1
incremental==17.5.0
leveldb==0.194
py==1.4.34
pycrypto==2.6.1
pytest==3.1.3
six==1.10.0
Twisted==17.5.0
zope.interface==4.4.2
